Predictions from a bagged model

The predict() function computes predictions from each of the models in the ensembles and returns a single aggregated value for each sample in new_data.

# S3 method for bagger
predict(object, new_data, type = NULL, ...)

Arguments

object
An object generated by bagger().

new_data
A data frame of predictors. If a recipe or formula were originally used, the original data should be passed here instead of a preprocessed version.

type
A single character value for the type of predictions. For regression models, type = 'numeric' is valid and 'class' and 'prob' are valid for classification models.

...
Not currently used.

Examples

data(airquality)

set.seed(7687)
rule_fit <- bagger(Ozone ~ ., data = airquality, model = "model_rules", B = 5)
predict(rule_fit, new_data = airquality[, -1])

#> # A tibble: 153 x 1
#>    .pred
#>    <dbl>
#>  1 22.0 
#>  2 21.1 
#>  3 20.1 
#>  4 21.2 
#>  5 11.9 
#>  6 12.1 
#>  7 22.3 
#>  8 13.1 
#>  9  8.59
#> 10 20.1 
#> # … with 143 more rows