Bagging functions

General suite of bagging functions for several models.

bagger(x, ...)

# S3 method for default
bagger(x, ...)

# S3 method for data.frame
bagger(x, y, model = "CART", B = 10L,
  opt = NULL, control = bag_control(), extract = NULL, ...)

# S3 method for matrix
bagger(x, y, model = "CART", B = 10L, opt = NULL,
  control = bag_control(), extract = NULL, ...)

# S3 method for formula
bagger(formula, data, model = "CART", B = 10L,
  opt = NULL, control = bag_control(), extract = NULL, ...)

# S3 method for recipe
bagger(x, data, model = "CART", B = 10L, opt = NULL,
  control = bag_control(), extract = NULL, ...)

Arguments

x
A data frame, matrix, or recipe (depending on the method being used).

...
Optional arguments to pass to the extract function.

y
A numeric or factor vector of outcomes. Categorical outcomes (i.e classes) should be represented as factors, not integers.

model
A single character value for the model being bagged. Possible values are "CART", "MARS", "C5.0" (classification only), and "model_rules" (regression only).

B
A single integer greater than 1 for the maximum number of bootstrap samples/ensemble members (some model fits might fail).

opt
A named list (or NULL) of arguments to pass to the underlying model function. A list of possible arguments per model are given in Details.

control
A list of options generated by bag_control().

extract
A function (or NULL) that can extract model-related aspects of each ensemble member. See Details and example below.

formula
An object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted. Note that this package does not support multivariate outcomes and that, if some predictors are factors, dummy variables will not be created unless by the underyling model function.

data
A data frame containing the variables used in the formula or recipe.

Details

bagger() fits separate models to bootstrap samples. The prediction function for each model object is encoded in an R expression and the original model object is discarded. When making predictions, each prediction formula is evaluated on the new data and aggregated using the mean.

Any arbitrary item can be saved from the model object (including the model object itself) using the extract argument, which should be a function with arguments x, and .... The results of this function are saved into a list column called extras (see the sample below).

Variable importance scores are calculated using implementations in each package. When requested, the results are in a tibble with column names term (the predictor), value (the importance score), and used (the percentage of times that the variable was in the prediction equation).

The models can be fit in parallel using the future package. The enable parallelism, use the future::plan() function to declare how the computations should be distributed. Note that this will almost certainly multiple the memory requirements to fit the models.

Examples

library(rsample)

#> Loading required package: tidyr

library(recipes)

#> Loading required package: dplyr

#> 
#> Attaching package: ‘dplyr’

#> The following object is masked from ‘package:testthat’:
#> 
#>     matches

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

#> 
#> Attaching package: ‘recipes’

#> The following object is masked from ‘package:stats’:
#> 
#>     step

library(dplyr)
library(yardstick)

#> For binary classification, the first factor level is assumed to be the event.
#> Set the global option `yardstick.event_first` to `FALSE` to change this.

data(biomass)

biomass_tr <-
  biomass %>%
  dplyr::filter(dataset == "Training") %>%
  dplyr::select(-dataset, -sample)

biomass_te <-
  biomass %>%
  dplyr::filter(dataset == "Testing") %>%
  dplyr::select(-dataset, -sample)

# ------------------------------------------------------------------------------

perf <- metric_set(rmse, rsq, yardstick::ccc)

ctrl <- bag_control(oob = perf, var_imp = TRUE)

# ------------------------------------------------------------------------------

# `B` is low to make the examples run faster

set.seed(7687)
mars_fit <- bagger(x = biomass_tr[, -6], y = biomass_tr$HHV,
                   model = "MARS", B = 5, control = ctrl)

#> 
#> Attaching package: ‘purrr’

#> The following object is masked from ‘package:testthat’:
#> 
#>     is_null

var_imp(mars_fit)

#> # A tibble: 5 x 3
#>   term      value  used
#>   <chr>     <dbl> <int>
#> 1 carbon   100        5
#> 2 hydrogen  19.7      5
#> 3 nitrogen  16.3      4
#> 4 oxygen    15.1      5
#> 5 sulfur     1.48     1

set.seed(7687)
cart_fit <- bagger(x = biomass_tr[, -6], y = biomass_tr$HHV,
                   model = "CART", B = 5, control = ctrl)

set.seed(7687)
rule_fit <- bagger(x = biomass_tr[, -6], y = biomass_tr$HHV,
                   model = "model_rules", B = 5, control = ctrl)

# ------------------------------------------------------------------------------
# Other interfaces

# Recipes can be used
biomass_rec <-
  recipe(HHV ~ ., data = biomass_tr) %>%
  step_pca(all_predictors())

set.seed(7687)
cart_pca_fit <- bagger(biomass_rec, data = biomass_tr, model = "CART", B = 5,
                       control = ctrl)

cart_pca_fit

#> Bagged CART (regression with 5 members)
#> 
#> Variable importance scores include:
#> 
#> # A tibble: 5 x 3
#>   term  value  used
#>   <chr> <dbl> <int>
#> 1 PC2   4500.     5
#> 2 PC1   3559.     5
#> 3 PC3   1107.     5
#> 4 PC5    648.     5
#> 5 PC4    468.     5
#> 

cart_pca_fit$oob

#> # A tibble: 3 x 4
#>   .metric  mean  stdev     n
#>   <chr>   <dbl>  <dbl> <int>
#> 1 ccc     0.880 0.0527     5
#> 2 rmse    1.71  0.178      5
#> 3 rsq     0.794 0.0961     5

cart_fit$oob

#> # A tibble: 3 x 4
#>   .metric  mean  stdev     n
#>   <chr>   <dbl>  <dbl> <int>
#> 1 ccc     0.906 0.0293     5
#> 2 rmse    1.53  0.190      5
#> 3 rsq     0.836 0.0532     5

# An example rule:
cart_pca_fit$model_df$.pred_form[[1]][[3]]

#> PC1 <= -52.3191755624344 & PC1 > -42.1720688864883 & PC1 <= -54.3142622814669 & 
#>     PC2 <= -19.7450992967057 ~ 14.4845625

# Using formulas
set.seed(7687)
rule_form<- bagger(HHV ~ ., data = biomass_tr, model = "model_rules", B = 5,
                   control = ctrl)

# ------------------------------------------------------------------------------
# Extracting model components

num_term_nodes <- function(x, ...) {
  tibble(num_nodes = sum(x$frame$var == "<leaf>"))
}

set.seed(7687)
with_extras <- bagger(x = biomass_tr[, -6], y = biomass_tr$HHV,
                   model = "CART", B = 5, extract = num_term_nodes)

dplyr::bind_rows(with_extras$model_df$extras)

#> # A tibble: 5 x 1
#>   num_nodes
#>       <int>
#> 1        35
#> 2        39
#> 3        40
#> 4        39
#> 5        39